Move rest of aggregation into apply
